+++ .github/workflows/test-core.yml
@@ -88,6 +88,7 @@ jobs:
             paths: >-
               tests/unit/bots/
               tests/unit/gateway/
+              tests/unit/cli/
             extra_ignore: ""
             pythonpath: >-
               ${{ github.workspace }}/src/praisonai-agents:${{ github.workspace }}/src/praisonai-code
