+++ src/praisonai-bot/praisonai_bot/scheduler/__init__.py
@@ -1,0 +1,5 @@
+"""Gateway scheduler bridge — agent dispatch for due scheduled jobs."""
+
+from .executor import JobResult, ScheduledAgentExecutor
+
+__all__ = ["ScheduledAgentExecutor", "JobResult"]
