+++ src/praisonai/tests/unit/test_c9_backward_compat.py
@@ -53,3 +53,12 @@ class TestBotModuleIdentity:
         old_mod = importlib.import_module("praisonai.integration.gateway_host")
         new_mod = importlib.import_module("praisonai_bot.integration.gateway_host")
         assert old_mod is new_mod
+
+    def test_scheduler_executor_shim(self):
+        from praisonai.scheduler.executor import ScheduledAgentExecutor as ShimExec
+        from praisonai_bot.scheduler.executor import ScheduledAgentExecutor as BotExec
+        from praisonai.scheduler.condition_gate import ShellConditionGate as ShimGate
+        from praisonai_bot.scheduler.condition_gate import ShellConditionGate as BotGate
+
+        assert ShimExec is BotExec
+        assert ShimGate is BotGate
